Reworking the compilation database, and moving their components into their own source file
